pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ipp"
version = "0.0.1"
authors = [
  "Thibault Vatter <thibault.vatter@gmail.com>" 
]
description = "Intro to Programming with Python"
readme = "README.md"
license = "MIT"
homepage = "https://github.com/tvatter/ipp"
repository = "https://github.com/tvatter/ipp.git"
issues = "https://github.com/tvatter/ipp/issues"
include = [
    { path = "tests", format = "sdist" },
    { path = "scripts", format = "sdist" }
]

[tool.poetry.dependencies]
python = "~=3.12"

[tool.poetry.group.dev.dependencies]
ruff = "*"
pytest = "*"
pyhamcrest = "*"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
testpaths = [
  "tests"
]